knn: backport item KNN type fix (#1121) to 2025

+++ src/lenskit/knn/item.py
@@ -1,6 +1,6 @@
 # This file is part of LensKit.
 # Copyright (C) 2018-2023 Boise State University.
-# Copyright (C) 2023-2025 Drexel University.
+# Copyright (C) 2023-2026 Drexel University.
 # Licensed under the MIT license, see LICENSE.md for details.
 # SPDX-License-Identifier: MIT
 
@@ -214,7 +214,7 @@ class ItemKNNScorer(Component[ItemList], Trainable):
         cmat = rmat / np.maximum(norms, np.finfo("f4").smallest_normal)
         assert cmat.shape == rmat.shape
         log.debug("[%s] normalized, memory use %s", timer, max_memory())
-        return cmat
+        return cmat.astype(np.float32)
 
     @override
     def __call__(self, query: QueryInput, items: ItemList) -> ItemList:
